Fix (some) visual studio warnings, related to narrowing conversions
Ok, one more fix.
The remaining visual studio warnings are either:
* Not directly in openMSX code.
* Not fixable without introducing warnings on the Linux side.
* In 'opll.cc' which we intentionally want to keep unchanged.
So for now I'm done with vs-warnings.

+++ src/utils/DivModBySame.hh
@@ -64,7 +64,7 @@ public:
 		assert(uint32_t(q) == q); // must fit in 32 bit
 		// result fits in 32-bit, so no 64-bit calculations required
 		uint32_t r = uint32_t(dividend) - uint32_t(q) * uint32_t(divisor);
-		return {q, r};
+		return {uint32_t(q), r};
 	}
 
 	[[nodiscard]] uint32_t mod(uint64_t dividend) const
